Moves down underline of links and fixes edges (#57)
The 'text-decoration-skip: edges' CSS property avoids the
display of a fraction of an underline, that looked like a dot,
after the closing parenthesis of the (PDF) suffixes added to
links to PDF/XLSX/DOCX documents.

The 'text-underline-position: under' had the same effect,
and also adds a unique touch to the visual identity of
all our sites.